store/requirements.txt
fastapi[all]==0.99.1
sqlalchemy==1.4.43
uvicorn[standard]==0.21.1
asyncio-redis==0.16.0
alembic==1.8.1
psycopg2-binary==2.9.5
pyjwt==2.6.0
celery==5.2.7
inflection==0.5.1
pytest==7.2.0
pytest-env==0.8.1
pre-commit==2.20.0
mock==4.0.3
Jinja2==3.1.2
email-validator==1.3.0
python-keycloak==2.16.2
pydantic-translations==0.2.1
l10n==0.1.1
redis==4.3.4
dogpile.cache==1.2.2
python-dotenv==0.21.0